# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dussehra Dolls API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  phone?: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  imageUrl?: string;
  stock: number;
  rating?: number;
  reviewCount?: number;
  isFeatured?: boolean;
  tags?: string[];
}

export interface CartItem {
  productId: number;
  product?: Product;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  subtotal: number;
}

export interface Order {
  id: number;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  address?: string;
  paymentMethod?: string;
  createdAt: string;
}

export interface CreateOrderRequest {
  address: string;
  paymentMethod: string;
}

export type ChitPlanStatus =
  (typeof ChitPlanStatus)[keyof typeof ChitPlanStatus];

export const ChitPlanStatus = {
  open: "open",
  active: "active",
  completed: "completed",
} as const;

export interface ChitPlan {
  id: number;
  name: string;
  description?: string;
  totalAmount: number;
  monthlyContribution: number;
  duration: number;
  members: number;
  maxMembers: number;
  startDate?: string;
  status: ChitPlanStatus;
}

export interface JoinChitRequest {
  chitPlanId: number;
}

export type ChitEnrollmentStatus =
  (typeof ChitEnrollmentStatus)[keyof typeof ChitEnrollmentStatus];

export const ChitEnrollmentStatus = {
  active: "active",
  completed: "completed",
  withdrawn: "withdrawn",
} as const;

export interface ChitEnrollment {
  id: number;
  chitPlan: ChitPlan;
  joinedAt: string;
  amountPaid: number;
  nextPaymentDate?: string;
  status: ChitEnrollmentStatus;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export interface Wallet {
  balance: number;
  transactions: Transaction[];
}

export interface AddFundsRequest {
  amount: number;
}

export type GetProductsParams = {
  category?: string;
  search?: string;
};