# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dussehra Dolls API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: products
    description: Product catalog
  - name: cart
    description: Shopping cart
  - name: orders
    description: Order management
  - name: chits
    description: Chit savings plans
  - name: wallet
    description: Wallet management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /chits:
    get:
      operationId: getChitPlans
      tags: [chits]
      summary: List available chit plans
      responses:
        "200":
          description: Chit plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChitPlan"
    post:
      operationId: joinChitPlan
      tags: [chits]
      summary: Join a chit plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinChitRequest"
      responses:
        "201":
          description: Joined chit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChitEnrollment"

  /chits/my:
    get:
      operationId: getMyChits
      tags: [chits]
      summary: Get user's chit enrollments
      responses:
        "200":
          description: My chits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChitEnrollment"

  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get wallet balance and transactions
      responses:
        "200":
          description: Wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
    post:
      operationId: addFunds
      tags: [wallet]
      summary: Add funds to wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFundsRequest"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
      required: [name, email, password]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [id, name, email]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required: [id, name, price, category, stock]

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]

    CartItem:
      type: object
      properties:
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        subtotal:
          type: number
      required: [productId, quantity, subtotal]

    AddToCartRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]

    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]

    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        address:
          type: string
        paymentMethod:
          type: string
        createdAt:
          type: string
      required: [id, status, items, total, createdAt]

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required: [productId, productName, quantity, price, subtotal]

    CreateOrderRequest:
      type: object
      properties:
        address:
          type: string
        paymentMethod:
          type: string
      required: [address, paymentMethod]

    ChitPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        totalAmount:
          type: number
        monthlyContribution:
          type: number
        duration:
          type: integer
        members:
          type: integer
        maxMembers:
          type: integer
        startDate:
          type: string
        status:
          type: string
          enum: [open, active, completed]
      required: [id, name, totalAmount, monthlyContribution, duration, members, maxMembers, status]

    JoinChitRequest:
      type: object
      properties:
        chitPlanId:
          type: integer
      required: [chitPlanId]

    ChitEnrollment:
      type: object
      properties:
        id:
          type: integer
        chitPlan:
          $ref: "#/components/schemas/ChitPlan"
        joinedAt:
          type: string
        amountPaid:
          type: number
        nextPaymentDate:
          type: string
        status:
          type: string
          enum: [active, completed, withdrawn]
      required: [id, chitPlan, joinedAt, amountPaid, status]

    Wallet:
      type: object
      properties:
        balance:
          type: number
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required: [balance, transactions]

    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
      required: [id, type, amount, description, createdAt]

    AddFundsRequest:
      type: object
      properties:
        amount:
          type: number
      required: [amount]